フロントエンド側のテーマシステム修正を完了
- Preferences画面のuseTheme依存を削除してwindow APIを使用
- ThemeProviderWrapperを完全に書き換えてイベントリスナーを実装
- 型定義をuseThemeStoreに統合してThemeContext.tsxを削除
- 全コンポーネントの型参照を新しい場所に更新
