+++ src/types/api/Window.ts
@@ -1,0 +1,3 @@
+export class WindowAPI {
+  static readonly NOTIFY_THEME_CHANGED = 'notify_theme_changed'
+}
